# server/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpacetimeWiki REST API",
    "description": "Programmatic CRUD and search for pages, collections, and more. SpacetimeWiki is a real-time wiki with rich text editing, hierarchical collections, tagging, search, SSO/OAuth/LDAP auth, SCIM provisioning, WebAuthn/passkeys, import/export, and more.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "description": "Full-text search with advanced filters.\n\nSupports query syntax:\n- `q` — free-text search against page title + content\n- `collection_id` — narrow to a specific collection\n- `author_id` — narrow to a specific author\n- `from` / `to` — date range on updated_at (ISO dates or epoch ms)\n- `tags` — tag:value filters, colon-separated, comma-delimited",
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Free-text search query",
              "title": "Q"
            },
            "description": "Free-text search query"
          },
          {
            "name": "collection_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by collection ID",
              "default": "",
              "title": "Collection Id"
            },
            "description": "Filter by collection ID"
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by author/user ID",
              "default": "",
              "title": "Author Id"
            },
            "description": "Filter by author/user ID"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Date range start (ms epoch or ISO date)",
              "default": "",
              "title": "From"
            },
            "description": "Date range start (ms epoch or ISO date)"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Date range end (ms epoch or ISO date)",
              "default": "",
              "title": "To"
            },
            "description": "Date range end (ms epoch or ISO date)"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated tag:value filters, e.g. 'status:published,priority:high'",
              "default": "",
              "title": "Tags"
            },
            "description": "Comma-separated tag:value filters, e.g. 'status:published,priority:high'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "description": "Max results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/autocomplete": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Autocomplete",
        "description": "Quick title-only autocomplete search.",
        "operationId": "autocomplete_api_v1_search_autocomplete_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Search query prefix",
              "title": "Q"
            },
            "description": "Search query prefix"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "description": "Max suggestions",
              "default": 10,
              "title": "Limit"
            },
            "description": "Max suggestions"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutocompleteResult"
                  },
                  "title": "Response Autocomplete Api V1 Search Autocomplete Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutocompleteResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Result ID",
            "default": ""
          },
          "page_id": {
            "type": "string",
            "title": "Page Id",
            "description": "Matched page ID",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Page title",
            "default": ""
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Page slug",
            "default": ""
          }
        },
        "type": "object",
        "title": "AutocompleteResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchFilters": {
        "properties": {
          "collection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id",
            "description": "Filter by collection ID"
          },
          "author_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Id",
            "description": "Filter by author/user ID"
          },
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From",
            "description": "Date range start"
          },
          "to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To",
            "description": "Date range end"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Tag filters"
          }
        },
        "type": "object",
        "title": "SearchFilters"
      },
      "SearchResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "type": "array",
            "title": "Data",
            "description": "Search results"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Original search query",
            "default": ""
          },
          "filters": {
            "$ref": "#/components/schemas/SearchFilters",
            "description": "Applied filters"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total result count",
            "default": 0
          }
        },
        "type": "object",
        "title": "SearchResponse"
      },
      "SearchResultItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Search result ID",
            "default": ""
          },
          "search_token": {
            "type": "string",
            "title": "Search Token",
            "description": "Search session token",
            "default": ""
          },
          "page_id": {
            "type": "string",
            "title": "Page Id",
            "description": "Matched page ID",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Matched page title",
            "default": ""
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Matched page slug",
            "default": ""
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt",
            "description": "Contextual excerpt with match highlighting",
            "default": ""
          },
          "match_type": {
            "type": "string",
            "title": "Match Type",
            "description": "Type of match: title, content, tag",
            "default": ""
          },
          "created_at": {
            "type": "integer",
            "title": "Created At",
            "description": "Sort timestamp",
            "default": 0
          }
        },
        "type": "object",
        "title": "SearchResultItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "SpacetimeWiki API server"
    }
  ]
}